# openapi.yaml
openapi: 3.1.0
info:
  title: API Contábil Universal
  version: 0.1.0
  description: |
    Hub contábil que unifica acesso a fontes oficiais (SERPRO Integra Contador,
    Focus NFe, LegisWeb, Open Finance) em um contrato único, interoperável com
    qualquer ERP ou sistema contábil.
  contact:
    name: Pedro
    email: pedromasterchif@gmail.com
servers:
  - url: https://api.contabil.exemplo.com/v1
    description: Produção
  - url: https://sandbox.api.contabil.exemplo.com/v1
    description: Sandbox
security:
  - OAuth2: [read, write]
...
paths:
...
  /simples-nacional/das:
    post:
      tags: [Simples Nacional]
      summary: Emitir DAS
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmitirDASRequest' }
      responses:
        '202':
          description: Em processamento (resultado via webhook)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperacaoAssincrona' }
        '200':
          description: Emitido sincronamente
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DAS' }
  /simples-nacional/pgdas/{cnpj}/{periodo}:
    get:
      tags: [Simples Nacional]
      summary: Consultar PGDAS-D
      parameters:
        - { name: cnpj, in: path, required: true, schema: { type: string } }
        - { name: periodo, in: path, required: true, schema: { type: string, example: '2026-03' } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PGDAS' }
...
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://auth.contabil.exemplo.com/oauth/token
          scopes:
            read: Leitura
            write: Escrita
  schemas:
...
    EmitirDASRequest:
      type: object
      required: [cnpj, periodo]
      properties:
        cnpj: { type: string }
        periodo: { type: string, example: '2026-03' }
        dataVencimento: { type: string, format: date }
    DAS:
      type: object
      properties:
        cnpj: { type: string }
        periodo: { type: string }
        valor: { type: number }
        codigoBarras: { type: string }
        pdfUrl: { type: string, format: uri }
        emitidoEm: { type: string, format: date-time }
    PGDAS:
      type: object
      properties:
        cnpj: { type: string }
        periodo: { type: string }
        receitaBruta: { type: number }
        valorTotal: { type: number }
        status: { type: string }
...
    OperacaoAssincrona:
      type: object
      properties:
        operacaoId: { type: string, format: uuid }
        status: { type: string, enum: [PENDENTE, PROCESSANDO, CONCLUIDA, FALHA] }
        webhookEvento: { type: string, example: 'das.emitida' }
...